rewrite: use bmxtranswrap
As wrapping with ffmpeg alone didn't create valid v1.3 op1a mxf in
ard zdf profile this has to be done with bmxtranswrap. This can swap
audio track or silence them. But it can't insert external audio. So if
external audio is provided do the swapping, muting and inserting with
ffmpeg and rewrap mxf with bmxtranswrap. Else do the swapping and mu-
ting with bmxtranswrap alone.

Signed-off-by: Conrad Zelck <git@simpel.cc>